Move Gem::Specification.bundled_stubs to bundler extension
