+++ lib/rubygems/specification.rb
@@ -833,19 +833,6 @@ class Gem::Specification < Gem::BasicSpecification
   end
 
   ##
-  # Returns a Gem::StubSpecification for bundled gems
-
-  def self.bundled_stubs
-    require_relative "../bundled_gems"
-    Gem.bundled_gems.map do |name|
-      Gem::Specification.find_by_name(name)
-    rescue Gem::MissingSpecError
-    end.compact
-  rescue LoadError
-    []
-  end
-
-  ##
   # Returns a Gem::StubSpecification for installed gem named +name+
   # only returns stubs that match Gem.platforms
